Ordenado los headers y separado la lógica en sus respectivas partes
