+++ CMakeLists.txt
@@ -3,5 +3,5 @@ project(EDA2_Lab C)
 
 set(CMAKE_C_STANDARD 11)
 
-add_executable(EDA2_Lab main.c sources/interface.c headers/interface.h sources/handleStack.c headers/handleStack.h headers/common.h sources/users.c headers/users.h)
+add_executable(EDA2_Lab main.c sources/interface.c headers/interface.h sources/handleStack.c headers/handleStack.h headers/common.h sources/users.c headers/users.h sources/user_arrays.c headers/user_arrays.h)
 set(CMAKE_EXE_LINKER_FLAGS "${CMAKE_EXE_LINKER_FLAGS} -mwindows -municode -mconsole")
