+++ headers/common.h
@@ -29,8 +29,6 @@
 #define TRUE 1
 #define FALSE 0
 
-
-
 // Forward declarations
 LRESULT CALLBACK WindowProc(HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam);
 LRESULT CALLBACK DialogProcedure(HWND hwnd, UINT msg, WPARAM wp, LPARAM lp);
