+++ sources/users.c
@@ -3,7 +3,6 @@
 //
 
 #include "../headers/users.h"
-#include "../headers/common.h"
 #include <time.h>
 #include <stdio.h>
 
@@ -26,7 +25,7 @@ void addUser(nodelist *nlist, user *newuser) {
     n->next = NULL;
 }
 
-user* findUser(nodelist nlist, wchar_t usernameid[]) {
+user* getUser(nodelist nlist, wchar_t usernameid[]) {
     wchar_t username[MAX_LENGTH];
     int id = -1;
     swscanf(usernameid, L"%[^#]#%d", username, &id);
@@ -42,59 +41,9 @@ user* findUser(nodelist nlist, wchar_t usernameid[]) {
 
 }
 
-void sendFriendRequest(user* active_user, user* friend) {
-    if (friend == NULL) {
-        return;
-    }
-    add_to_array(&active_user->friend_requests_sent, friend);
-    add_to_array(&friend->friend_requests_received, active_user);
-}
-
-void acceptFriendRequest(user* active_user, user* friend) {
-    if (friend == NULL) {
-        return;
-    }
-    remove_from_array(&active_user->friend_requests_received, friend);
-    remove_from_array(&friend->friend_requests_sent, active_user);
-    add_to_array(&active_user->friends, friend);
-    add_to_array(&friend->friends, active_user);
-}
-
-void rejectFriendRequest(user* active_user, user* friend) {
-    if (friend == NULL) {
-        return;
-    }
-    remove_from_array(&active_user->friend_requests_received, friend);
-    remove_from_array(&friend->friend_requests_sent, active_user);
-}
 
 
-void add_to_array(user_array* array, user* addeduser) {
-    if (array->size == 0) {
-        array->users = malloc(sizeof(user*));
-    } else {
-        array->users = realloc(array->users, sizeof(user*) * (array->size + 1));
-    }
-
-    array->users[array->size] = addeduser;
-    array->size += 1;
-}
-
-void remove_from_array(user_array* array, user* removeduser) {
-    int removed = FALSE;
-    for (int i = 0; i < array->size; i++){
-        if (removed == TRUE){
-            array->users[i-1] = array->users[i];
-        }
-        if (array->users[i]->id == removeduser->id){
-            removed = TRUE;
-        }
-    }
-
-    array->size -= 1;
-}
-
-void OperateAs(HWND hwnd, nodelist users) {
+void operateAs(HWND hwnd, nodelist users) {
     HWND windowH = CreateWindowExW(0, L"DialogWindow", L"Choose a user", WS_VISIBLE | WS_OVERLAPPEDWINDOW,
                                    CW_USEDEFAULT, CW_USEDEFAULT, 248, 480, hwnd, NULL, NULL, NULL);
 
@@ -114,16 +63,7 @@ void OperateAs(HWND hwnd, nodelist users) {
     }
 }
 
-int search_User(user_array users, int id){
-    for (int i = 0; i < users.size; i++){
-        if (id == users.users[i]->id){
-            return TRUE;
-        }
-    }
-    return FALSE;
-}
-
-void SendFRModal(HWND hwnd, nodelist users, user active_user) {
+void sendFRModal(HWND hwnd, nodelist users, user active_user) {
     HWND windowH = CreateWindowExW(0, L"DialogWindow", L"Enviar solicitud de amistad", WS_VISIBLE | WS_OVERLAPPEDWINDOW,
                                    CW_USEDEFAULT, CW_USEDEFAULT, 248, 480, hwnd, NULL, NULL, NULL);
